+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Use a worker so the UI doesn't freeze when rendering big graphs.
 - MUCH smaller WebAssembly module.
 - Added `--package` flag to open a specific published package.
+- Added a `hide-isolated` filter to remove unpublished or root workspace
+  packages with no incoming dependencies.
 
 ## 6.0.0